# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Koza encryption API",
    "version": "1.0.0",
    "description": "Stateless AES-256-GCM / JWE adapter. Requests are processed on the server; the browser app encrypts locally. Each encryption uses a fresh random key and IV. Keep keys separate from encrypted output."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/v1/encrypt": {
      "post": {
        "summary": "Encrypt text or bytes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncryptInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Encrypted"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decrypt": {
      "post": {
        "summary": "Decrypt and authenticate a package",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecryptInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Decrypted"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server-side API access token, separate from per-package decryption keys."
      }
    },
    "schemas": {
      "EncryptInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "text": {
            "type": "string"
          },
          "data": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]*$",
            "description": "Canonical unpadded base64url bytes; maximum decoded length 10 MiB."
          },
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "mime": {
            "type": "string",
            "maxLength": 255
          },
          "sculpture": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$",
            "description": "Optional public sculpture-frame SHA-256 digest. No additional entropy is credited. The API does not render a sculpture."
          }
        },
        "oneOf": [
          {
            "required": [
              "text"
            ],
            "not": {
              "required": [
                "data"
              ]
            }
          },
          {
            "required": [
              "data"
            ],
            "not": {
              "required": [
                "text"
              ]
            }
          }
        ]
      },
      "DecryptInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jwe",
          "key"
        ],
        "properties": {
          "jwe": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "description": "KOZA1-, legacy ESIK1- or raw canonical base64url 256-bit key."
          }
        }
      },
      "Encrypted": {
        "type": "object",
        "required": [
          "jwe",
          "key",
          "algorithm"
        ],
        "properties": {
          "jwe": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "algorithm": {
            "const": "A256GCM"
          },
          "sculpture": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$",
            "description": "Optional public sculpture-frame SHA-256 digest. No additional entropy is credited. The API does not render a sculpture."
          }
        }
      },
      "Decrypted": {
        "type": "object",
        "required": [
          "kind",
          "name",
          "mime"
        ],
        "properties": {
          "kind": {
            "enum": [
              "text",
              "file"
            ]
          },
          "name": {
            "type": "string"
          },
          "mime": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "data": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]*$",
            "description": "Canonical unpadded base64url bytes; maximum decoded length 10 MiB."
          },
          "sculpture": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$",
            "description": "Optional public sculpture-frame SHA-256 digest. No additional entropy is credited. The API does not render a sculpture."
          }
        },
        "oneOf": [
          {
            "properties": {
              "kind": {
                "const": "text"
              }
            },
            "required": [
              "text"
            ]
          },
          {
            "properties": {
              "kind": {
                "const": "file"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}